Update aesthetics
Applied CSS to all webpages. Index has some internal CSS since it should look a little different from the other pages. The TJ bear picture is ID'd so that many unique styles can be applied to it. Some style properties were googled online and are not yet a part of the odin curriculim.